images: move 2 forms to Views/ (+ test for ImageOriginalLink)
Per #4331. Two forms move; two stay reusable on closer inspection.

- `Components::CommercialInquiryForm` →
  `Views::Controllers::Images::Emails::Form`. Used only by
  `Images::EmailsController` (`#new` HTML + turbo_stream ModalTurboForm).
- `Components::ImageVoteAnonymityForm` →
  `Views::Controllers::Images::Votes::Anonymity::Form`. Used only by
  `Images::Votes::AnonymityController#edit`.

Staying in `app/components/` (audit was imprecise — Phlex kit
invocations don't show up in `Components::<Name>` greps):

- `ImageOriginalLink` — rendered from `images/show` AND
  `Components::LightboxCaption` (the latter is reusable, so the link
  has to be too). Added a previously-missing component test
  (`test/components/image_original_link_test.rb`) to lock its
  contract — the component was at 100% coverage via downstream views
  but lacked a dedicated test.
- `ImageVoteInterface` — rendered from `images/show` (ImagesController)
  AND `images/votes/_update.erb` (Images::VotesController turbo_stream
  response). Two controllers → reusable, same call as
  `OccurrenceProjectsForm`.

Cleanup-in-scope (fix-on-touch per testing rule):

- `image_vote_anonymity_form_test.rb`: replaced two forbidden
  `assert_html(html, "body", text: ...)` + `assert_includes(html, "5")`
  patterns with `<p strong>` text extraction via Nokogiri.
- `commercial_inquiry_form_test.rb`: replaced
  `assert_html(html, "body", text: label)` + `assert_includes(html, message)`
  with `label[for='email_message']` + `textarea[name='email[message]']`
  selectors carrying `text:`.

54 controller tests + 25 component/view tests pass, lint clean.
