+++ app/views/controllers/images/emails/new.erb
@@ -1,6 +1,6 @@
 <% add_page_title(:commercial_inquiry_title.t(name: @image.unique_format_name)) %>
 
-<%= render(Components::CommercialInquiryForm.new(
+<%= render(Views::Controllers::Images::Emails::Form.new(
       image: @image,
       user: @user,
       message: @message,
