+++ app/views/controllers/images/votes/anonymity/edit.html.erb
@@ -1,6 +1,6 @@
 <% add_page_title(:image_vote_anonymity_title.t) %>
 
-<%= render(Components::ImageVoteAnonymityForm.new(
+<%= render(Views::Controllers::Images::Votes::Anonymity::Form.new(
       FormObject::ImageVoteAnonymity.new(
         num_anonymous: @num_anonymous,
         num_public: @num_public
